backend/requirements.txt
asgiref==3.6.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
config==0.5.1
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2
django-cors-headers==4.0.0
django-filter==23.5
django-redis==5.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
urllib3==2.1.0
